+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.3.0`.
+- Updated [@adguard/tsurlfilter] to `v3.5.0`.
+
 ## [3.2.15] - 2025-11-12
 
 ### Changed